Added RIPEMD-160 and RIPEMD-128 implementations.
Also added "fcrypt_memzero" with some checks for OS-specific functions in
configure.ac. Uses an OS-provided explicit_memzero function if possible.
If that isn't avaliable then it uses a memory barrier with clang or gcc.
Uses a volatile pointer to memset to zero the memory as a last resort if
none of those are possible.

Windows provides SecureMemzero(ptr, len). I think you can check for
windows by using "#ifdef _WIN32" but I'm not sure what header it is in.